implicit

+++ Implicit.java
@@ -1,8 +1,8 @@
-
+// byte > short > int > long > float > double
 public class Implicit {
     public static void main(String[] args) {
         double x = 2.0;
-        int y = (int)x + 2;
+        double y = x + 2;
         System.out.println(y);
 
     }
